測試助教測資 & Revise Assignment 7
加入助教在midterm2前釋出的測資 & parser的warning消失

+++ src/shape_parser.h
@@ -101,7 +101,7 @@ public:
     {
         int num_index = 0; //第幾個數字
         std::string temp;
-        for (int i = 0; i < rule.size(); i++)
+        for (size_t i = 0; i < rule.size(); i++)
         {
             temp = scanner.nextToken();
             if (rule.at(i) == "num")
